feat(embedding): hybrid subject+body vector search with weighted cosine (#205)
* feat(embedding): hybrid subject+body vector search with weighted cosine

Semantic search previously embedded a single concatenated string per message,
which meant location (postcode and area) dominated similarity — two sofas in
the same town clustered tighter than a sofa and a sofa-bed across the country.
Split into two vectors per message (subject-only, body-only) and combine with
0.7*subject + 0.3*body cosine, keeping the existing MinVectorScore threshold
and whole-word keyword boost.

- messages_embeddings.embedding renamed to subject_embedding; body_embedding
  BLOB NULL added; table truncated so the first re-embed rebuilds both.
- Subject preprocessing reuses the existing bracket-counting "TYPE: item
  (location)" parser from IncomingMailService — extracted to SubjectParser
  so mail and embedding pipelines cannot drift. No regex / postcode guessing.
- Body embedded raw (no regex location stripping — high risk of cutting real
  descriptions).
- New embeddings:regenerate command for a full wipe-and-rebuild of live
  messages (messages_spatial where successful=0 AND promised=0).
- Go embedding.Store.Load reads both columns; Entry gains SubjectVec + BodyVec
  (nullable); Search returns weighted combined score (body_cos=0 when absent).
- vectorsearch keeps the whole-word keyword boost from #198 on the combined
  score.

Fixes cache leak between tests: Cache::flush() added to the base TestCase
setUp so rate-limit entries (e.g. bounce_autoreply:<hash>) don't survive
across tests. DatabaseTransactions isolates the DB; the cache was the gap.

Tests:
- 1506/1506 Go
- 1633/1633 Laravel
- SubjectParser pins nested-bracket location handling
- EmbeddingService pins canonical-subject parse and empty-body NULL
- MessagesEmbeddingsSchema pins the migration shape


* test(go): run all package tests, not just ./test/...

Status API and CI were only running the integration tests under
./test/..., so package-level tests in embedding, donations, utils, etc.
were invisible to CI. Switching to ./... picks up 242 additional tests
(1532 → 1774) with no other changes needed.

The donations package had a latent build error (%d format verb on a
utils.FlexFloat64) that the old pattern masked — fixed alongside.


---------

+++ iznik-server-go/donations/stripe.go
@@ -184,7 +184,7 @@ func CreateSubscription(c *fiber.Ctx) error {
 
 	// Create a Stripe product for this subscription.
 	prodParams := &stripe.ProductParams{
-		Name: stripe.String(fmt.Sprintf("Freegle Monthly Donation - £%d", req.Amount)),
+		Name: stripe.String(fmt.Sprintf("Freegle Monthly Donation - £%.2f", float64(req.Amount))),
 	}
 	prod, err := product.New(prodParams)
 	if err != nil {
